Add color-scheme meta tag
