+++ my-portfolio/src/hooks/useTheme.js
@@ -1,0 +1,24 @@
+import { useState, useEffect } from 'react';
+
+export const useTheme = () => {
+  const [darkMode, setDarkMode] = useState(() => {
+    const stored = localStorage.getItem('theme');
+    if (stored) {
+      return stored === 'dark';
+    }
+    return window.matchMedia('(prefers-color-scheme: dark)').matches;
+  });
+
+  useEffect(() => {
+    const classList = document.documentElement.classList;
+    if (darkMode) {
+      classList.add('dark');
+      localStorage.setItem('theme', 'dark');
+    } else {
+      classList.remove('dark');
+      localStorage.setItem('theme', 'light');
+    }
+  }, [darkMode]);
+
+  return [darkMode, setDarkMode];
+};
